+++ distribution/uniform/doc.go
@@ -1,0 +1,14 @@
+// Package uniform implements a uniform continuous random variable
+//
+// Uniform distribution is a probability distribution that has constant
+// probability.
+//
+// The distribution is parameterized by two parameters, namely, a and b,
+// which are its minimum and maximum values, respectively.
+//
+//
+// References:
+//
+// https://en.wikipedia.org/wiki/Uniform_distribution_(continuous)
+// http://mathworld.wolfram.com/UniformDistribution.html
+package uniform